requirements.txt
rasa==2.8.2
rasa-sdk==2.8.1
streamlit
cryptography==3.4.8
beautifulsoup4==4.12.1
openai==0.27.0
ruamel.yaml==0.16.13
pyyaml==6.0.1
yamlfix==1.11.0
langchain==0.0.27
Flask==2.1.2
Flask-PyMongo==2.3.0
rasa==2.8.2
regex==2021.7.6
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
sanic==20.12.6
Sanic-Cors==0.10.0.post3
sanic-jwt==1.5.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.2
scipy==1.7.3
sklearn-crfsuite==0.3.6
slackclient==2.9.4
SQLAlchemy==1.4.36
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.4
tensorflow-addons==0.13.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.4
tensorflow-gpu-estimator==2.3.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-probability==0.13.0
tf-estimator-nightly==2.9.0.dev2022012709
tqdm==4.64.0
transformers==3.0.0
requests
requests-oauthlib
requests-toolbelt
urllib3
typing_extensions